requirements.txt
torch
torchvision
numpy
opencv-python
Pillow
scikit-learn
tqdm
matplotlib
streamlit
